aviso de cierre o abierto
